arreglos e integracion a los ultimos cambios(csv -> json)
