+++ src/main/java/org/example/querys/QueryElementsFromZero.java
@@ -5,6 +5,8 @@ import org.example.models.Library;
 import org.example.models.Recipe;
 
 import java.util.List;
+import java.util.Set;
+import java.util.HashSet;
 
 public class QueryElementsFromZero implements Query {
 	@Override
@@ -18,19 +20,16 @@ public class QueryElementsFromZero implements Query {
 		    }
 		}
 
-
         if (recipe != null) {
             System.out.println("\nArbol de ingredientes de\n" + element.name() + ":");
 
             processRecipe(recipe, libraries, "    ");
             // Usar ElementCounter para mostrar el conteo de elementos básicos
-            ElementCounter counter = new ElementCounter();
-            counter.countAllBasicElements(recipe, libraries);
+            countAllBasicElements(recipe, libraries);
             return recipe;
         }
 
-        System.out.println("No se encontró receta para: " + element.name() + ":(");
-        System.out.println("--------------------------------------------------------------");
+        System.out.println("No se encontró receta para: " + element.name());
         return null;
     }
 
@@ -53,11 +52,67 @@ public class QueryElementsFromZero implements Query {
 
             System.out.println(indent + (isLast ? "└── " : "├── ") + ingredient.name());
             
-            
+            // Si el ingrediente es a su vez una receta, procesamos recursivamente
             if (craftableIngredient != null) {
                 String nextIndent = indent + (isLast ? "    " : "│   ");
                 processRecipe(craftableIngredient, libraries, nextIndent);
             }
         }
+    }
+
+    //Cuenta y muestra el total de cada elemento basico que hay en una receta
+	public void countAllBasicElements(Recipe recipe, List<Library> libraries) {
+		Set<String> basicElements = getBasicElements(libraries);
+		
+		System.out.println("\nElementos básicos necesarios:");
+		for (String elementName : basicElements) {
+			long count = countBasicElement(recipe, elementName, libraries);
+			if (count > 0) {
+				System.out.println("  " + elementName + ": " + count);
+			}
+		}
+	}
+    
+	//Cuenta todos los elementos de un tipo en una receta
+	//Se le envia una receta, un elemento basico como H , y la lista de recetas
+    public long countBasicElement(Recipe recipe, String elementName, List<Library> libraries) {
+        long count = 0;
+
+        for (Element ingredient : recipe.ingredients()) {
+            if (ingredient.name().equals(elementName)) {
+                count++;
+            } else {
+            	Recipe nestedRecipe = null;
+            	for (Library l : libraries) {
+            	    if (l.recipe().give().name().equals(ingredient.name())) {
+            	        nestedRecipe = l.recipe();
+            	        break;
+            	    }
+            	}
+
+                if (nestedRecipe != null) {
+                    count += countBasicElement(nestedRecipe, elementName, libraries);
+                }
+            }
+        }
+
+        return count;
+    }
+  
+    //Arma un set de todos los elementos basicos que hay en una receta
+    private Set<String> getBasicElements(List<Library> libraries) {
+        Set<String> allElements = new HashSet<>();
+        Set<String> recipeResults = new HashSet<>();
+
+        for (Library library : libraries) {
+            Recipe recipe = library.recipe();
+            recipeResults.add(recipe.give().name());
+            for (Element element : recipe.ingredients()) {
+                allElements.add(element.name());
+            }
+        }
+
+        allElements.removeAll(recipeResults);
+        return allElements;
     }
 }
