+++ src/test/resources/recipes.json
@@ -110,5 +110,5 @@
     "time": 5.0,
     "type": "liquido",
     "elements": ["agua","dioxido_carbono"]
-  }
+  } 
 ]
